* Fixed bug in CemiLSerializer * Fixed setting priority in ControlField1 * Added Routing Send functionality * Added automatic initialization of KnxNetIPFrameSerializerFactory

+++ src/Tiveria.Home.Knx/Cemi/Serializers/CemiLDataSerializer.cs
@@ -84,7 +84,7 @@ namespace Tiveria.Home.Knx.Cemi.Serializers
             writer.Write((byte)cemiMessage.ControlField2.RawData);
             cemiMessage.SourceAddress.Write(writer);
             cemiMessage.DestinationAddress.Write(writer);
-            writer.Write(cemiMessage.Apci.Size - 1); // Apci structure includes TPCI and therefore Size is NPDULength+1
+            writer.Write((byte)(cemiMessage.Apci.Size - 1)); // Apci structure includes TPCI and therefore Size is NPDULength+1
             cemiMessage.Apci.Write(writer);
         }
     }
